Fixed: Deleting answers ajax (task7-ajax)

+++ app/controllers/answers_controller.rb
@@ -20,12 +20,8 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    if current_user.author?(answer)
-      answer.destroy
-      redirect_to answer.question, notice: 'Your answer has been deleted.'
-    else
-      redirect_to answer.question, notice: 'You are not the author of the answer.'
-    end
+    @id = answer.id
+    answer.destroy if current_user.author?(answer)
   end
 
   private
